Minor hint fixes.
Removed an extraneous "They say" from the Ocarina of Time location hint. Commented out the Fairy Ocarina and Ocarina of Time item hints as they are no longer items that are placed. Also added a missing period to a junk hint and fixed the spacing of the line marker comment. Some minor code clean-up as well.